# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: drive
    description: Google Drive operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /drive/upload:
    post:
      operationId: uploadDocumentToDrive
      tags: [drive]
      summary: Upload document to Google Drive
      description: Creates a Google Doc in Drive with the document data and returns a shareable URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriveUploadRequest"
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriveUploadResponse"
        "503":
          description: Google Drive not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Upload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /drive/files/{fileId}/preview:
    get:
      operationId: getDriveFilePreview
      tags: [drive]
      summary: Stream a Drive file preview
      description: Returns an inline image stream for a Drive file ID so the web app can render previews without relying on third-party cookies.
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Preview image stream
          content:
            image/*:
              schema:
                type: string
                format: binary
        "400":
          description: Missing file ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Google Drive not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Preview failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DriveUploadRequest:
      type: object
      properties:
        documentId:
          type: string
          description: Firestore document ID
        title:
          type: string
        studentName:
          type: string
        school:
          type: string
        serviceType:
          type: string
        academicYear:
          type: string
        agent:
          type: string
        date:
          type: string
        templateName:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
        notes:
          type: string
      required:
        - documentId
        - title
        - studentName
        - serviceType

    DriveUploadResponse:
      type: object
      properties:
        fileId:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
      required:
        - fileId
        - fileUrl
        - fileName

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error